Remove inconsistent gaiacli keys new/add dualism (#2904)
* Remove inconsistent gaiacli keys new/add dualism
