+++ client/keys/root.go
@@ -20,7 +20,6 @@ func Commands() *cobra.Command {
 	}
 	cmd.AddCommand(
 		mnemonicKeyCommand(),
-		newKeyCommand(),
 		addKeyCommand(),
 		listKeysCmd,
 		showKeysCmd(),
